fastapi_server/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
langchain
langchain-core
langgraph
openai
async-timeout
websockets
httpx
pandas
langchain-experimental
tabulate
scikit-learn
